server/pyproject.toml
[project]
name = "openai-voice-agents-sdk-sample"
version = "0.1.0"
description = "Example python backend server for the Agents SDK"
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "numpy>=2.2.3",
    "openai>=1.68.2",
    "openai-agents[voice]>=0.0.6",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
